Moved component to its own folder.
