+++ slider/src/app/appBase.cpp
@@ -2,12 +2,12 @@
 
 void Slider::AppBase::SetupComponents()
 {
-    for (auto& component : m_Components)
+    for (const auto& component : m_Components)
         component->Setup();
 }
 
 void Slider::AppBase::UpdateComponents()
 {
-    for (auto& component : m_Components)
+    for (const auto& component : m_Components)
         component->Update();
 }
