+++ slider/src/component/component.cpp
